cpuid: limit search to current values (#8)
Improved the APIs to make it clearer when leafs are not present. There
are a few ways to identify when to stop scanning mentioned in the
manual. For displaying facts, the code will now not only use a zero
check in some cases, It will also pay attention to the advertized limits
from the CPU.

Fixed item found by lint